+++ src/jcpsantiago/arqivist/api/confluence/handlers.clj
@@ -167,7 +167,7 @@
         (-> (sql/find-by-keys db-connection :slack_teams {:atlassian_tenant_id tenant_id})
             first)]
 
-    ;; drop the row corresponding to the current tenant, plus the row in  'slack_teams'
+    ;; drop the row corresponding to the current tenant, plus the row in 'slack_teams'
     (mulog/with-context {:slack-team-id external_team_id :slack-team-name team_name}
                         (try
                           (sql/delete! db-connection :atlassian_tenants {:id tenant_id})
@@ -183,12 +183,13 @@
                             ;; if there's a slack team connected, we have to uninstall the app
                             (let [res (-> @(httpkit/get
                                             (str "https://slack.com/api/apps.uninstall?"
-                                                 "client_id=" (get-in system [:env :slack :arqivist-slack-client-id])
-                                                 "&client_secret=" (get-in system [:env :slack :arqivist-slack-client-secret]))
+                                                 "client_id=" (get-in system [:slack-env :client-id])
+                                                 "&client_secret=" (get-in system [:slack-env :client-secret]))
                                             {:headers {"Content-Type" "application/json; charset=utf-8"}
                                              :oauth-token access_token})
                                           :body
                                           (jsonista/read-value jsonista/keyword-keys-object-mapper))]
+
                               (cond
                                 (spec/invalid? (spec/conform ::slack-specs/apps-uninstall res))
                                 (do
@@ -215,7 +216,8 @@
                                   (-> {:status 500 :body "Couldn't uninstall from Slack!"} (content-type "text-plain"))))))
 
                           (catch Exception e
-                            ;; TODO: send a Slack message to the admin user informing them this failed, and they need to manually remove the app
+                            ;; TODO: send a Slack message to the admin user informing them this failed,
+                            ;; and they need to manually remove the app
                             (mulog/log ::uninstalling-app
                                        :success :false
                                        :error (.getMessage e)
@@ -232,7 +234,7 @@
           event-type (:eventType lifecycle-payload)
           base-url (:baseUrl lifecycle-payload)
 
-          {:keys [:atlassian_tenants/tenant_id] :as atlassian_tenant}
+          {:keys [:atlassian_tenants/id] :as atlassian_tenant}
           (-> (sql/find-by-keys
                db-connection
                :atlassian_tenants
@@ -241,7 +243,7 @@
 
       (mulog/log ::lifecycle-event :event-type event-type :base-url base-url :local-time (java.time.LocalDateTime/now))
 
-      (mulog/with-context {:base-url (:baseUrl lifecycle-payload) :event-type event-type :tenant-id tenant_id}
+      (mulog/with-context {:base-url (:baseUrl lifecycle-payload) :event-type event-type :tenant-id id}
                           (case event-type
                             "installed" (installed lifecycle-payload atlassian_tenant system)
                             "enabled" (enabled lifecycle-payload atlassian_tenant system)